requirements.txt
asgiref==3.8.1
crispy-bootstrap5==2024.2
Django==4.2.13
django-autoslug==1.9.9
django-crispy-forms==2.2
importlib_metadata==8.0.0
Markdown==3.6
pillow==10.3.0
psycopg2==2.9.9
sqlparse==0.5.0
typing_extensions==4.12.2
zipp==3.19.2
